Add a script allowing to set harbor configuration for integration tests
`ansible-test` is ignoring environment variables (on purpose), so we can not
use environment variable to dynamically set the harbor url/credentials.

We can still set the values by using the roles `vars` directories that have precedence
on the roles defaults.

The script added in this commit allows to retrieve the credentials from environment
and create the `vars/main.yml` for all the roles present in the `tests/integration/targets/`.